Create fs, server daemon packages for testing
